requirements.txt
Django>=4.1.0,<4.2
djangorestframework>=3.14.0,< 3.15
drf-spectacular>=0.25.0,<0.26
mysqlclient>=2.1.1,<2.2
Pillow>=9.4.0,<9.5